// files-manager-web/src/api/generated.ts
export class ActionDto implements IActionDto {
    type?: ActionTypeDto;
    target?: string | undefined;

    constructor(data?: IActionDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.type = _data["type"];
            this.target = _data["target"];
        }
    }

    static fromJS(data: any): ActionDto {
        data = typeof data === 'object' ? data : {};
        let result = new ActionDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["type"] = this.type;
        data["target"] = this.target;
        return data;
    }
}

export interface IActionDto {
    type?: ActionTypeDto;
    target?: string | undefined;
}

export enum ActionTypeDto {
    Move = "Move",
    Copy = "Copy",
    Delete = "Delete",
    Rename = "Rename",
}

export class ConditionDto implements IConditionDto {
    type?: ConditionTypeDto;
    value?: string | undefined;

    constructor(data?: IConditionDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.type = _data["type"];
            this.value = _data["value"];
        }
    }

    static fromJS(data: any): ConditionDto {
        data = typeof data === 'object' ? data : {};
        let result = new ConditionDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["type"] = this.type;
        data["value"] = this.value;
        return data;
    }
}

export interface IConditionDto {
    type?: ConditionTypeDto;
    value?: string | undefined;
}

export enum ConditionTypeDto {
    Extension = "Extension",
}
// ...
export interface IRuleDto {
    id?: string;
    name?: string | undefined;
    description?: string | undefined;
    enabled?: boolean;
    priority?: number;
    conditions?: ConditionDto[] | undefined;
    actions?: ActionDto[] | undefined;
}